fix(ios): mission monitoring fixes, worker-count consistency, polish + ControlView split
Bugs found by driving the app in the simulator against a prod boss
mission with 16 workers:

- Blank frozen conversation (intermittent on cold start and after
  layout changes): LazyVStack + .defaultScrollAnchor(.bottom) anchored
  the viewport past estimated content. Conversations now render in a
  plain VStack under 80 rows (ConditionallyLazyVStack), with a rescue
  scrollTo on appear and a per-mission .id() remount of the scroll view.
- Chat scrolled through the transparent nav bar (text over the status
  bar/title): explicit opaque toolbarBackground.
- Mission switcher sheet material seam: explicit presentationBackground.
- 'WebSocket failed to open' transport errors rendered as chat bubbles:
  filtered by reason; WS attempts stop after 3 consecutive open
  failures (prod closes /api/control/ws with 1011 every time).
- False 'Slow connection · catching up' banner on healthy quiet
  streams: degraded now requires a failed health probe, not staleness.
- Worker counts disagreed between pill (1 active) and sheet (6 active):
  shared WorkerBuckets (active/waiting/done/failed) drives both, with
  a new Waiting section + summary card in the sheet.

Polish (mirrors the web dashboard's surface-sheen treatment): top-edge
light catch + edge highlight on composer, bubbles, pills and cards;
gradient user bubble; assistant bubble border now matches its uneven
corner shape.

Refactor: ControlView.swift 6945 → ~4500 lines; bubbles, switcher sheet
and support views extracted to MessageBubbles.swift,
MissionSwitcherSheet.swift, ControlSupportViews.swift (mechanical
split, no behavior change).
