+++ ios_dashboard/SandboxedDashboard/Services/APIService.swift
@@ -838,6 +838,14 @@ final class APIService {
 
     // MARK: - Control Streaming
 
+    /// After this many consecutive WebSocket open failures, stop trying WS
+    /// for the rest of the process lifetime and go straight to SSE. Some
+    /// deployments (reverse proxies without upgrade support) close every WS
+    /// attempt with 1011; retrying on each reconnect just adds latency and
+    /// noise before the inevitable SSE fallback.
+    nonisolated static let webSocketMaxOpenFailures = 3
+    private var webSocketOpenFailureStreak = 0
+
     func streamControl(
         missionId: String?,
         sinceSeq: Int64? = nil,
@@ -850,7 +858,7 @@ final class APIService {
 
         return Task { [weak self] in
             guard let self else { return }
-            if preferWebSocket, token != nil {
+            if preferWebSocket, token != nil, self.webSocketOpenFailureStreak < Self.webSocketMaxOpenFailures {
                 let opened = await self.runControlWebSocket(
                     missionId: missionId,
                     sinceSeq: sinceSeq,
@@ -859,9 +867,13 @@ final class APIService {
                     onDiagnostic: onDiagnostic,
                     onEvent: onEvent
                 )
+                if opened {
+                    self.webSocketOpenFailureStreak = 0
+                }
                 if opened || Task.isCancelled {
                     return
                 }
+                self.webSocketOpenFailureStreak += 1
                 self.emitDiagnostic(
                     transport: .webSocket,
                     phase: .fallback,
